+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add get_json_with_retries to json_client to permit retrying requests.
+
 # 59.2.1
 
 * Warn when `GdsApi::Rummager` is initialised and `GdsApi::TestHelpersRummager`